Changed Organization and Updated CSS
Changed the organization of files. Removed Header.js because all content that would have been in that component will be in the Nav.js component. Added HeroSection.js, Highlights.js, Testimonials.js, and About.js, which will be a part of Main.js.

Updated the general CSS in index.html. Updated specific CSS in Nav.js.